requirements.txt
streamlit
langgraph
langchain
langchain-core
langchain-community
pydantic
python-dotenv
faiss-cpu
pandas
numpy
sqlalchemy
sentence-transformers
pypdf
tqdm
anthropic
pytest
